// src/gen-ts/src/models/Company.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { AuditUser } from './AuditUser'
import {
  AuditUserFromJSON,
  AuditUserFromJSONTyped,
  AuditUserToJSON,
  AuditUserToJSONTyped,
} from './AuditUser'
import type { CompanyType } from './CompanyType'
import {
  CompanyTypeFromJSON,
  CompanyTypeFromJSONTyped,
  CompanyTypeToJSON,
  CompanyTypeToJSONTyped,
} from './CompanyType'

/**
 *
 * @export
 * @interface Company
 */
export interface Company {
  /**
   *
   * @type {string}
   * @memberof Company
   */
  comment?: string
  /**
   *
   * @type {Date}
   * @memberof Company
   */
  created_at?: Date
  /**
   *
   * @type {Date}
   * @memberof Company
   */
  updated_at?: Date
  /**
   *
   * @type {AuditUser}
   * @memberof Company
   */
  created_by?: AuditUser
  /**
   *
   * @type {AuditUser}
   * @memberof Company
   */
  updated_by?: AuditUser
  /**
   *
   * @type {string}
   * @memberof Company
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof Company
   */
  name?: string
  /**
   *
   * @type {string}
   * @memberof Company
   */
  rib?: string
  /**
   *
   * @type {string}
   * @memberof Company
   */
  description?: string
  /**
   *
   * @type {CompanyType}
   * @memberof Company
   */
  company_type?: CompanyType
}

/**
 * Check if a given object implements the Company interface.
 */
export function instanceOfCompany(value: object): value is Company {
  return true
}

export function CompanyFromJSON(json: any): Company {
  return CompanyFromJSONTyped(json, false)
}

export function CompanyFromJSONTyped(json: any, ignoreDiscriminator: boolean): Company {
  if (json == null) {
    return json
  }
  return {
    comment: json['comment'] == null ? undefined : json['comment'],
    created_at: json['created_at'] == null ? undefined : new Date(json['created_at']),
    updated_at: json['updated_at'] == null ? undefined : new Date(json['updated_at']),
    created_by: json['created_by'] == null ? undefined : AuditUserFromJSON(json['created_by']),
    updated_by: json['updated_by'] == null ? undefined : AuditUserFromJSON(json['updated_by']),
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name'],
    rib: json['rib'] == null ? undefined : json['rib'],
    description: json['description'] == null ? undefined : json['description'],
    company_type:
      json['company_type'] == null ? undefined : CompanyTypeFromJSON(json['company_type']),
  }
}

export function CompanyToJSON(json: any): Company {
  return CompanyToJSONTyped(json, false)
}

export function CompanyToJSONTyped(
  value?: Company | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    comment: value['comment'],
    created_at:
      value['created_at'] == null ? value['created_at'] : value['created_at'].toISOString(),
    updated_at:
      value['updated_at'] == null ? value['updated_at'] : value['updated_at'].toISOString(),
    created_by: AuditUserToJSON(value['created_by']),
    updated_by: AuditUserToJSON(value['updated_by']),
    id: value['id'],
    name: value['name'],
    rib: value['rib'],
    description: value['description'],
    company_type: CompanyTypeToJSON(value['company_type']),
  }
}