// src/gen-ts/src/models/CrupdateCompany.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { CompanyType } from './CompanyType'
import {
  CompanyTypeFromJSON,
  CompanyTypeFromJSONTyped,
  CompanyTypeToJSON,
  CompanyTypeToJSONTyped,
} from './CompanyType'

/**
 *
 * @export
 * @interface CrupdateCompany
 */
export interface CrupdateCompany {
  /**
   *
   * @type {string}
   * @memberof CrupdateCompany
   */
  comment?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateCompany
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateCompany
   */
  name?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateCompany
   */
  rib?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateCompany
   */
  description?: string
  /**
   *
   * @type {CompanyType}
   * @memberof CrupdateCompany
   */
  company_type?: CompanyType
}

/**
 * Check if a given object implements the CrupdateCompany interface.
 */
export function instanceOfCrupdateCompany(value: object): value is CrupdateCompany {
  return true
}

export function CrupdateCompanyFromJSON(json: any): CrupdateCompany {
  return CrupdateCompanyFromJSONTyped(json, false)
}

export function CrupdateCompanyFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CrupdateCompany {
  if (json == null) {
    return json
  }
  return {
    comment: json['comment'] == null ? undefined : json['comment'],
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name'],
    rib: json['rib'] == null ? undefined : json['rib'],
    description: json['description'] == null ? undefined : json['description'],
    company_type:
      json['company_type'] == null ? undefined : CompanyTypeFromJSON(json['company_type']),
  }
}

export function CrupdateCompanyToJSON(json: any): CrupdateCompany {
  return CrupdateCompanyToJSONTyped(json, false)
}

export function CrupdateCompanyToJSONTyped(
  value?: CrupdateCompany | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    comment: value['comment'],
    id: value['id'],
    name: value['name'],
    rib: value['rib'],
    description: value['description'],
    company_type: CompanyTypeToJSON(value['company_type']),
  }
}