// src/gen-ts/src/models/CompanyType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const CompanyType = {
    BTP: 'BTP',
    HOTEL: 'HOTEL'
} as const;
export type CompanyType = typeof CompanyType[keyof typeof CompanyType];


export function instanceOfCompanyType(value: any): boolean {
    for (const key in CompanyType) {
        if (Object.prototype.hasOwnProperty.call(CompanyType, key)) {
            if (CompanyType[key as keyof typeof CompanyType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function CompanyTypeFromJSON(json: any): CompanyType {
    return CompanyTypeFromJSONTyped(json, false);
}

export function CompanyTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): CompanyType {
    return json as CompanyType;
}

export function CompanyTypeToJSON(value?: CompanyType | null): any {
    return value as any;
}

export function CompanyTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): CompanyType {
    return value as CompanyType;
}